projects/2025/Selenium/team_8/requirements.txt
pandas==2.2.3
python-dotenv==1.1.1
requests==2.32.3
selenium==4.29.0
chromedriver-autoinstaller==0.10.4
